Pull out data prep for components and signal components into a helper
