+++ moped-editor/src/views/projects/projectView/ProjectComponents/utils/useCreateComponent.js
@@ -1,15 +1,9 @@
 import { useReducer } from "react";
 import { useMutation } from "@apollo/client";
 import { ADD_PROJECT_COMPONENT } from "src/queries/components";
-import {
-  makeDrawnLinesInsertionData,
-  makeDrawnPointsInsertionData,
-  makeLineStringFeatureInsertionData,
-  makePointFeatureInsertionData,
-} from "./makeFeatures";
+import { makeComponentInsertData } from "./makeComponentData";
 import { getDrawId, isDrawnDraftFeature } from "./features";
 import { removeFeatureFromDraftComponent } from "./onMapClick";
-import { knackSignalRecordToFeatureSignalsRecord } from "src/utils/signalComponentHelpers";
 
 const createReducer = (state, action) => {
   switch (action.type) {
@@ -143,80 +137,12 @@ export const useCreateComponent = ({
     setLinkMode(null);
   };
 
-  const createComponentInsertData = (component) => {
-    /* Start data preparation */
-    const {
-      component_id,
-      description,
-      moped_subcomponents,
-      component_name,
-      internal_table,
-      features,
-    } = component || createState.draftComponent;
-
-    const subcomponentsArray = moped_subcomponents
-      ? moped_subcomponents.map((subcomponent) => ({
-          subcomponent_id: subcomponent.value,
-        }))
-      : [];
-
-    const featureTable = internal_table;
-
-    const featuresToInsert = [];
-    const signalFeaturesToInsert = [];
-    const drawnLinesToInsert = [];
-    const drawnPointsToInsert = [];
-
-    const drawnFeatures = features.filter((feature) =>
-      isDrawnDraftFeature(feature)
-    );
-    const selectedFeatures = features.filter(
-      (feature) => !isDrawnDraftFeature(feature)
-    );
-
-    if (featureTable === "feature_street_segments") {
-      makeLineStringFeatureInsertionData(
-        featureTable,
-        selectedFeatures,
-        featuresToInsert
-      );
-      makeDrawnLinesInsertionData(drawnFeatures, drawnLinesToInsert);
-    } else if (featureTable === "feature_intersections") {
-      makePointFeatureInsertionData(
-        featureTable,
-        selectedFeatures,
-        featuresToInsert
-      );
-      makeDrawnPointsInsertionData(drawnFeatures, drawnPointsToInsert);
-    } else if (featureTable === "feature_signals") {
-      features.forEach((feature) => {
-        const signalRecord = knackSignalRecordToFeatureSignalsRecord(feature);
-        signalFeaturesToInsert.push(signalRecord);
-      });
-    }
-
-    return {
-      description,
-      component_id,
-      name: component_name,
-      project_id: projectId,
-      moped_proj_components_subcomponents: {
-        data: subcomponentsArray,
-      },
-      [featureTable]: {
-        data: featuresToInsert,
-      },
-      feature_drawn_lines: { data: drawnLinesToInsert },
-      feature_drawn_points: { data: drawnPointsToInsert },
-      feature_signals: { data: signalFeaturesToInsert },
-    };
-  };
-
   /**
    * Prepare component and feature data for component creation and call mutation/reset state
    */
   const onSaveDraftComponent = () => {
-    const newComponentData = createComponentInsertData(
+    const newComponentData = makeComponentInsertData(
+      projectId,
       createState.draftComponent
     );
 
@@ -236,7 +162,10 @@ export const useCreateComponent = ({
    * Prepare signal component and feature data for component creation and call mutation/reset state
    */
   const onSaveDraftSignalComponent = (signalComponent) => {
-    const newComponentData = createComponentInsertData(signalComponent);
+    const newComponentData = makeComponentInsertData(
+      projectId,
+      signalComponent
+    );
 
     addProjectComponent({ variables: { object: newComponentData } })
       .then(() => {
